+++ fnshare2/dosfunc.h
@@ -1,10 +1,10 @@
 #ifndef _DOSFUNC_H
 #define _DOSFUNC_H
 
-#include "types.h"
+#include "dosdata.h"
 
-extern ulong dos_ftime(void);
+extern uint32_t dos_ftime(void);
 extern void set_sft_owner(SFTREC_PTR sft);
-extern int is_a_character_device(uint dos_ds);
+extern int is_a_character_device(uint16_t dos_ds);
 
 #endif /* _DOSFUNC_H */
